move canInsertNode util to core
